Corrected a lot of styling, the header now works 90% still needs work to correct the z-index of the online display and to adjust the alignment of the logo/title. Split views directory into admin/basic to provide a better structure. System is now fully validated and client side is built to recieve the messages. Pages are now responsive however work still needs done to properly align forms. Security and error handling has been tested and is fully functioning. This should be one of the last commits before final scrubbing / commenting / tidying of code

+++ models/taskDao.js
@@ -98,12 +98,18 @@ class TaskDao {
    //Get the item to delete from the database
    const doc = await this.find(querySpec) //Execute the query
 
-   //Delete the fetched item
-   const { resource: failed } = await this.container
-     .item(doc[0].id, querySpec.parameters[0].value) //PartitionKey is extracted from querySpec
-     .delete(); //Delete the item
+   //Check if item exists
+   if(doc.length){ //Item exists
 
-   return !failed; //Informs controller of success/failure of operation
+     //Delete the fetched item
+     const { resource: failed } = await this.container
+       .item(doc[0].id, querySpec.parameters[0].value) //PartitionKey is extracted from querySpec
+       .delete(); //Delete the item
+
+     return !failed; //Informs controller of success/failure of operation
+ }else{ //Item does not exist
+   return false; //Cannot be deleted
+ }
  } //End deleteItem()
 } //End TaskDao
